Cambia el tipo de dato del script para la columna version.

+++ scripts/issue336.sql
@@ -1,4 +1,4 @@
 ALTER TABLE producto
-ADD version int(11) default 1 after urlImagen;
+ADD version bigint(11) default 1 after urlImagen;
 ALTER TABLE cantidadensucursal
-ADD version int(11) default 1 after estante;
+ADD version bigint(11) default 1 after estante;
